+++ CHANGES.rst
@@ -17,6 +17,10 @@ Deprecated
 Features added
 --------------
 
+* Add :meth:`~sphinx.application.Sphinx.add_static_dir` for copying static
+  assets from extensions to the build output.
+  Patch by Jared Dillard
+
 Bugs fixed
 ----------